requirements.txt
pandas
numpy
pandas-profiling
geopandas
pandas_gbq
db_dtypes
lxml
requests
BeautifulSoup4
webdriver_manager
selenium
matplotlib
seaborn
plotly
folium
shapely
pingouin
statsmodels
scikit-learn
torch 
tensorflow
opencv-python 
albumentations 
efficientnet-pytorch==0.7.1
tqdm
streamlit
streamlit-option-menu
streamlit_pandas_profiling
jupyterlab
google-cloud-bigquery
google-cloud-bigquery[pandas]
googlemaps